refactor v2 eval op step tests
